Moringa-school-work/Phase-5/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
pytest==7.4.3